Workaround for web-image.
