+++ web-image/src/com.oracle.svm.hosted.webimage/src/com/oracle/svm/hosted/webimage/wasm/codegen/WebImageWasmCodeGen.java
@@ -35,23 +35,18 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.TreeMap;
 
 import com.oracle.svm.core.BuildArtifacts;
-import com.oracle.svm.core.BuildPhaseProvider;
 import com.oracle.svm.core.InvalidMethodPointerHandler;
 import com.oracle.svm.core.image.ImageHeapLayoutInfo;
 import com.oracle.svm.core.meta.MethodPointer;
 import com.oracle.svm.hosted.HeapBreakdownProvider;
 import com.oracle.svm.hosted.NativeImageGenerator;
 import com.oracle.svm.hosted.image.AbstractImage;
-import com.oracle.svm.hosted.image.NativeImageHeap;
 import com.oracle.svm.hosted.meta.HostedMethod;
 import com.oracle.svm.hosted.webimage.WebImageCodeCache;
 import com.oracle.svm.hosted.webimage.WebImageHostedConfiguration;
@@ -336,19 +331,6 @@ public abstract class WebImageWasmCodeGen extends WebImageCodeGen {
      */
     protected void validateModule() {
         new WasmValidator().visitModule(module);
-    }
-
-    protected void afterHeapLayout() {
-        // after this point, the layout is final and must not be changed anymore
-        assert !hasDuplicatedObjects(codeCache.nativeImageHeap) : "heap.getObjects() must not contain any duplicates";
-        BuildPhaseProvider.markHeapLayoutFinished();
-        codeCache.nativeImageHeap.getLayouter().afterLayout(codeCache.nativeImageHeap);
-    }
-
-    protected boolean hasDuplicatedObjects(NativeImageHeap heap) {
-        Set<NativeImageHeap.ObjectInfo> deduplicated = Collections.newSetFromMap(new IdentityHashMap<>());
-        deduplicated.addAll(heap.getObjects());
-        return deduplicated.size() != heap.getObjectCount();
     }
 
     @Override
